Retry to receive bucets from db
When started at boot, it is possible the call to receive buckets is done
before influxdb is ready. If it is not ready no update is then posted to
the db later. Then by adding retry for this an application can try
several times to get the bucket it need. Then it can start to post
updates when the db is ready instead of end up in a failed state.